app/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.4
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
